Prepare zstd-complete 0.1.0 release

+++ ruzstd/src/encoding/levels/c_port/compress_bound.rs
@@ -1,10 +1,10 @@
 //! Worst-case compressed-size bound matching `ZSTD_COMPRESSBOUND()`.
 
 const ZSTD_BLOCKSIZE_MAX: usize = 128 * 1024;
-const ZSTD_MAX_INPUT_SIZE_64: usize = 0xFF00_FF00_FF00_FF00;
+const ZSTD_MAX_INPUT_SIZE_64: u64 = 0xFF00_FF00_FF00_FF00;
 
 pub(crate) fn compress_bound(src_size: usize) -> usize {
-    if usize::BITS == 64 && src_size >= ZSTD_MAX_INPUT_SIZE_64 {
+    if usize::BITS == 64 && (src_size as u64) >= ZSTD_MAX_INPUT_SIZE_64 {
         return 0;
     }
